safari-534.53-branch/LayoutTests/canvas/philip/tests/2d.text.font.parse.complex.html

<!DOCTYPE html>
<title>Canvas test: 2d.text.font.parse.complex</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.font = 'small-caps italic 400 12px/2 Unknown Font, sans-serif';
_assertSame(ctx.font, 'italic small-caps 12px "Unknown Font", sans-serif', "ctx.font", "'italic small-caps 12px \"Unknown Font\", sans-serif'");


});
</script>
